FP_2022/Lab1/requirements.txt
argparse~=1.4.0
google-api-python-client~=2.37.0
google-auth-httplib2~=0.1.0
google-auth-oauthlib~=0.4.6
geopy~=2.2.0
haversine~=2.5.1
firebase-admin~=5.2.0